# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenMapX API",
    "description": "Generated from the API's own route table and the integration route registry. Do not edit by hand — run `pnpm openapi:generate`.",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/admin/extensions/catalog/{id}": {
      "get": {
        "operationId": "getApiAdminExtensionsCatalogId",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "tags": [
          "admin"
        ],
        "x-openmapx-auth": "admin"
      }
    },
...
    "/api/admin/extensions/update/{id}": {
      "post": {
        "operationId": "postApiAdminExtensionsUpdateId",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "tags": [
          "admin"
        ],
        "x-openmapx-auth": "admin"
      }
    },
    "/api/admin/extensions/{id}": {
      "delete": {
        "operationId": "deleteApiAdminExtensionsId",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "tags": [
          "admin"
        ],
        "x-openmapx-auth": "admin"
      }
    },
...
    "/api/admin/integrations/{integrationId}/bindings": {
      "get": {
        "operationId": "getApiAdminIntegrationsIntegrationIdBindings",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "integrationId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "tags": [
          "admin"
        ],
        "x-openmapx-auth": "admin"
      }
    },
    "/api/admin/integrations/{integrationId}/bindings/{capability}": {
      "post": {
        "operationId": "postApiAdminIntegrationsIntegrationIdBindingsCapability",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "integrationId",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "capability",
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "serviceId"
                ],
                "properties": {
                  "serviceId": {
                    "type": "string",
                    "minLength": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "tags": [
          "admin"
        ],
        "x-openmapx-auth": "admin"
      },
      "delete": {
        "operationId": "deleteApiAdminIntegrationsIntegrationIdBindingsCapability",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "integrationId",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "capability",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "tags": [
          "admin"
        ],
        "x-openmapx-auth": "admin"
      }
    },
...
    "/api/attribution/motis-file/{filename}": {
      "get": {
        "operationId": "getApiAttributionMotisFileFilename",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "filename",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "tags": [
          "attribution"
        ],
        "x-openmapx-auth": "public"
      }
    },
    "/api/attribution/{sourceId}": {
      "get": {
        "operationId": "getApiAttributionSourceId",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "sourceId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "tags": [
          "attribution"
        ],
        "x-openmapx-auth": "public"
      }
    },
...
  }
}